refactor(worker): migrate dispatch ownership into a dedicated subtree
Move the authored-down, release, and observer dispatch paths out of the
flat worker module into worker/dispatch/ with a shared dispatch/mod.rs
and a dedicated timing.rs. Redistribute dispatch functions across the
subtree so every dispatch file stays under its 700-line (250-line for
mod.rs) hard cap and every dispatch function under 180 lines.

Split oversize functions in place rather than allowlisting:
- publisher_down_send_outcome -> resolve_slo_terminal_step (observer.rs)
- admit_authored_down -> finalize_allowed_admission (authored.rs)
- interpret_down_send_timing -> resolve_send_boundaries (timing.rs)
- prepare_authored_batch_view relocated to timing.rs

Tighten the architecture checker so oversized dispatch functions are hard
errors (not warnings) and assert the per-file caps in the regression test
module_line_limits_strictly_respected. Legacy downs.rs, down_outcome.rs,
and releases.rs paths are deleted.
