+++ rust/crates/sky_player_rs/src/engine/worker/dispatch/authored.rs
@@ -1,66 +1,23 @@
-use super::super::{
+use super::super::super::{
     ActionKind, DurationTicks, HARD_LATE_ABORT_THRESHOLD_US, LatencyClass, PlaybackClockState,
     QpcClock, QpcTicks, RuntimeDispatchCoordinator, SendLatencyEstimator, SharedMetrics,
     TRACE_KIND_DOWN, TRACE_KIND_UP, TelemetryCollector, TimelineTicks, TrackedKeyState,
     try_publish_metrics,
 };
-use super::down_outcome::{
-    commit_suppressed_up_request, interpret_down_send_timing, publisher_down_send_outcome,
-    read_qpc_us, record_blocked_unfocused_telemetry,
+use super::super::{
+    DownAdmission, WorkerConfig, WorkerHealthState, WorkerMetricsLocal, WorkerResources,
+    WorkerRuntime, WorkerTimingState, build_dispatch_budget, ensure_preflight_for_target,
+    final_down_admission, focus_matches, load_target_stamp, suspend_live_input,
+    target_stamp_still_current,
 };
-use super::{
-    DispatchPath, DownAdmission, WorkerConfig, WorkerHealthState, WorkerMetricsLocal,
-    WorkerResources, WorkerRuntime, WorkerTimingState, build_dispatch_budget,
-    ensure_preflight_for_target, final_down_admission, focus_matches, load_target_stamp,
-    planning::NextDispatchPlan, suspend_live_input, target_stamp_still_current,
+use super::observer::{
+    commit_suppressed_up_request, publisher_down_send_outcome, record_blocked_unfocused_telemetry,
 };
+use super::timing::{interpret_down_send_timing, prepare_authored_batch_view, read_qpc_us};
+use super::{AuthoredBatchView, AuthoredPacketContext, DispatchStep};
 use crate::engine::telemetry::TRACE_KIND_MIXED;
-use sky_dispatch_core::coordinator::PreparedBatch;
-use sky_dispatch_core::model::ScanCodeBatch;
 use smallvec::SmallVec;
 use std::sync::atomic::{AtomicBool, AtomicIsize, AtomicU64, Ordering};
-
-pub(crate) enum DispatchStep {
-    NoWork,
-    Dispatched,
-    Continue,
-    Terminate(String),
-}
-
-pub(crate) struct AuthoredPacketContext<'a> {
-    pub(crate) dispatch_plan: &'a NextDispatchPlan,
-    pub(crate) effective_now_ticks: TimelineTicks,
-    pub(crate) now_ticks: QpcTicks,
-    pub(crate) latency_class: LatencyClass,
-    pub(crate) focus_loss_fault: bool,
-}
-
-/// Snapshot of the prepared authored batch plus the projection of the
-/// schedule view used by admission, send, and telemetry.
-///
-/// Built once per authored epoch by `prepare_authored_batch_view`; the
-/// send/admission/telemetry helpers consume it without re-querying the
-/// coordinator schedule.
-pub(super) struct AuthoredBatchView {
-    pub(super) prepared_batch: PreparedBatch,
-    pub(super) batch_source_action_index: u32,
-    pub(super) batch_intent_count: usize,
-    pub(super) batch_kind: ActionKind,
-    pub(super) batch_scheduled_ticks: TimelineTicks,
-    pub(super) batch_scheduled_us: u64,
-    pub(super) authored_batch_scheduled_ticks: TimelineTicks,
-    pub(super) authored_batch_scheduled_us: u64,
-    pub(super) conflict_mask: u16,
-    pub(super) dispatch_path: DispatchPath,
-    pub(super) packet_mode: bool,
-    pub(super) packet_masks: Option<sky_dispatch_win32::input::PhysicalPacket>,
-    pub(super) scan_batch: ScanCodeBatch,
-}
-
-/// `Err(None)` indicates an unrecoverable terminal step; `Ok(None)` means the
-/// coordinator offered no authored work for this epoch (worker should advance
-/// the wait deadline instead).
-type BatchViewResult = Result<Option<AuthoredBatchView>, DispatchStep>;
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn dispatch_authored_packet(
@@ -198,154 +155,6 @@ pub(crate) fn dispatch_authored_packet(
         effective_now_ticks,
         lead_down_ticks,
     )
-}
-
-/// Project the prepared authored batch into a snapshot used by admission, send,
-/// and telemetry.  Built once per epoch so the worker does not re-query the
-/// coordinator schedule across multiple invariants within a single loop epoch
-/// (D1 — one immutable dispatch plan per epoch).
-fn prepare_authored_batch_view(
-    coordinator: &mut RuntimeDispatchCoordinator,
-    qpc_clock: QpcClock,
-    prepared_batch: PreparedBatch,
-) -> BatchViewResult {
-    let batch_index = prepared_batch.index;
-    let batch_scheduled_ticks = prepared_batch.effective_scheduled_ticks;
-    let packet_mode = match prepared_batch.packet_kind {
-        Some(sky_dispatch_core::model::PhysicalPacketKind::DownOnly)
-            if prepared_batch.packet_batch_count == 1 =>
-        {
-            false
-        }
-        Some(_) => true,
-        None => false,
-    };
-    let (
-        batch_kind,
-        dispatch_path,
-        batch_source_action_index,
-        batch_intent_count,
-        conflict_mask,
-        scan_batch,
-        packet_masks,
-    ) = if packet_mode {
-        let packet_view = match coordinator
-            .schedule
-            .view_packet_ticks(prepared_batch.packet_index, batch_scheduled_ticks)
-        {
-            Ok(value) => value,
-            Err(error) => {
-                return Err(DispatchStep::Terminate(format!(
-                    "runtime packet view failure: {error}"
-                )));
-            }
-        };
-        let conflict_mask = coordinator
-            .check_packet_down_conflicts(packet_view.up_mask(), packet_view.down_intents);
-        let up_count = packet_view.up_mask().count_ones() as usize;
-        let down_count = packet_view.down_mask().count_ones() as usize;
-        let dispatch_path = match prepared_batch.packet_kind {
-            Some(sky_dispatch_core::model::PhysicalPacketKind::UpOnly) => {
-                DispatchPath::UpOnly { up_count }
-            }
-            Some(sky_dispatch_core::model::PhysicalPacketKind::DownOnly) => {
-                DispatchPath::DownOnly { down_count }
-            }
-            Some(sky_dispatch_core::model::PhysicalPacketKind::Mixed) => DispatchPath::Mixed {
-                up_count,
-                down_count,
-            },
-            None => DispatchPath::DownOnly { down_count: 0 },
-        };
-        (
-            if matches!(dispatch_path, DispatchPath::UpOnly { .. }) {
-                ActionKind::Up
-            } else {
-                ActionKind::Down
-            },
-            dispatch_path,
-            packet_view
-                .header
-                .down_source_action_index
-                .or_else(|| {
-                    coordinator
-                        .schedule
-                        .batches
-                        .get(packet_view.header.first_batch_index as usize)
-                        .map(|batch| batch.source_action_index)
-                })
-                .unwrap_or(0),
-            up_count + down_count,
-            conflict_mask,
-            packet_view.down_scan_code_batch(),
-            Some(sky_dispatch_win32::input::PhysicalPacket::new(
-                packet_view.up_mask(),
-                packet_view.down_mask(),
-            )),
-        )
-    } else {
-        let batch_view = match coordinator
-            .schedule
-            .view_batch_ticks(batch_index, batch_scheduled_ticks)
-        {
-            Ok(value) => value,
-            Err(error) => {
-                return Err(DispatchStep::Terminate(format!(
-                    "runtime schedule view failure: {error}"
-                )));
-            }
-        };
-        let conflict_mask = coordinator.check_down_conflicts_compact(batch_view.intents);
-        (
-            batch_view.kind(),
-            match batch_view.kind() {
-                ActionKind::Up => DispatchPath::UpOnly {
-                    up_count: batch_view.intents.len(),
-                },
-                ActionKind::Down => DispatchPath::DownOnly {
-                    down_count: batch_view.intents.len(),
-                },
-            },
-            batch_view.source_action_index(),
-            batch_view.intents.len(),
-            conflict_mask,
-            batch_view.scan_code_batch_excluding_mask(conflict_mask),
-            None,
-        )
-    };
-    let batch_scheduled_us = match qpc_clock.timeline_to_us(batch_scheduled_ticks) {
-        Ok(value) => value,
-        Err(error) => {
-            return Err(DispatchStep::Terminate(format!(
-                "schedule telemetry conversion failure: {error:?}"
-            )));
-        }
-    };
-    let authored_batch_scheduled_ticks = coordinator.batch_scheduled_ticks[batch_index];
-    let authored_batch_scheduled_us = match qpc_clock.timeline_to_us(authored_batch_scheduled_ticks)
-    {
-        Ok(value) => value,
-        Err(error) => {
-            return Err(DispatchStep::Terminate(format!(
-                "authored schedule telemetry conversion failure: {error:?}"
-            )));
-        }
-    };
-    Ok(Some(AuthoredBatchView {
-        prepared_batch,
-        batch_source_action_index,
-        batch_intent_count,
-        batch_kind,
-        batch_scheduled_ticks,
-        batch_scheduled_us,
-        authored_batch_scheduled_ticks,
-        authored_batch_scheduled_us,
-        conflict_mask,
-        dispatch_path,
-        packet_mode,
-        packet_masks,
-        scan_batch,
-    }))
 }
 
 /// Admission gate + SendInput call + telemetry + estimator update + health
@@ -618,25 +427,44 @@ fn admit_authored_down(
                 return Ok(AdmissionOutcome::TargetChanged);
             }
         }
-        let frozen_budget = build_dispatch_budget(
+        return Ok(finalize_allowed_admission(
+            view,
+            config,
+            health,
             estimator,
-            view.dispatch_path,
             latency_class,
-            health.options,
-            config.timing.strict_timing,
-        );
-        let trace_kind = match view.prepared_batch.packet_kind {
-            Some(sky_dispatch_core::model::PhysicalPacketKind::UpOnly) => TRACE_KIND_UP,
-            Some(sky_dispatch_core::model::PhysicalPacketKind::DownOnly) => TRACE_KIND_DOWN,
-            Some(sky_dispatch_core::model::PhysicalPacketKind::Mixed) => TRACE_KIND_MIXED,
-            None => TRACE_KIND_DOWN,
-        };
-        return Ok(AdmissionOutcome::Allowed {
-            frozen_budget,
-            trace_kind,
-        });
+        ));
     }
     Ok(AdmissionOutcome::ConflictReject)
+}
+
+/// Contstructs the allowed admission from the frozen dispatch budget and the
+/// packet trace kind.  Extracted to keep `admit_authored_down` under the per
+/// dispatch-function line limit.
+fn finalize_allowed_admission(
+    view: &AuthoredBatchView,
+    config: &WorkerConfig,
+    health: &WorkerHealthState,
+    estimator: &SendLatencyEstimator,
+    latency_class: LatencyClass,
+) -> AdmissionOutcome {
+    let frozen_budget = build_dispatch_budget(
+        estimator,
+        view.dispatch_path,
+        latency_class,
+        health.options,
+        config.timing.strict_timing,
+    );
+    let trace_kind = match view.prepared_batch.packet_kind {
+        Some(sky_dispatch_core::model::PhysicalPacketKind::UpOnly) => TRACE_KIND_UP,
+        Some(sky_dispatch_core::model::PhysicalPacketKind::DownOnly) => TRACE_KIND_DOWN,
+        Some(sky_dispatch_core::model::PhysicalPacketKind::Mixed) => TRACE_KIND_MIXED,
+        None => TRACE_KIND_DOWN,
+    };
+    AdmissionOutcome::Allowed {
+        frozen_budget,
+        trace_kind,
+    }
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -677,7 +505,6 @@ fn record_down_send_outcome(
     if let Some(error) = backend.timing_error.take() {
         return DispatchStep::Terminate(format!("QPC failure after note-on: {error:?}"));
     }
-    let trace_kind = *trace_kind;
     let result_success = result.is_success();
     let result_started_ticks = result.evidence.started_ticks;
     let result_completed_ticks = result.evidence.completed_ticks;
@@ -696,6 +523,7 @@ fn record_down_send_outcome(
             view.batch_source_action_index
         ));
     }
+    let trace_kind = *trace_kind;
     finalize_down_send_outcome(
         view,
         config,
